requirements.txt
PyAutoGUI>=0.9.52
Pillow>=8.2.0
openpyxl>=3.0.7